refactor: consolidate session token summary into token_summary.py Per reviewer feedback, moved session aggregation logic into token_summary.py instead of a separate module. Removed session_token_summary.py. Single cmd_token_summary handler now serves both commands.
