+++ apptrace/src/monocle_apptrace/token_summary.py
@@ -165,3 +165,87 @@ def format_table(rows):
         lines.append(fmt.format(*[str(row[k]) for k in col_keys]))
     lines.append(sep)
     return "\n".join(lines)
+
+# ---------------------------------------------------------------------------
+# Per-session summary
+# ---------------------------------------------------------------------------
+
+SESSION_ATTR = "scope.agentic.session"
+
+_SESSION_HEADERS = ["Session", "Model", "Input", "Cache Read", "Cache Create", "Output", "Total"]
+
+
+def summarize_by_session(time_window="all", monocle_dir=None):
+    # type: (str, Optional[Path]) -> List[Dict]
+    """Aggregate token usage per session from trace files in *monocle_dir*."""
+    if monocle_dir is None:
+        monocle_dir = MONOCLE_DIR
+
+    cutoff = _window_cutoff(time_window)
+    aggregated: dict = defaultdict(lambda: defaultdict(lambda: defaultdict(int)))
+
+    if not monocle_dir.exists():
+        return []
+
+    for trace_file in sorted(monocle_dir.glob("monocle_trace_*.json")):
+        ts = _parse_timestamp_from_filename(trace_file.name)
+        if ts is None:
+            continue
+        if cutoff is not None and ts < cutoff:
+            continue
+
+        for span in _load_spans(trace_file):
+            attrs = span.get("attributes", {})
+            session_id = attrs.get(SESSION_ATTR)
+            model = attrs.get("entity.2.name")
+            if not session_id or not model:
+                continue
+            for event in span.get("events", []):
+                if event.get("name") != "metadata":
+                    continue
+                event_attrs = event.get("attributes", {})
+                for key in _TOKEN_KEYS:
+                    val = event_attrs.get(key, 0)
+                    if isinstance(val, (int, float)):
+                        aggregated[session_id][model][key] += int(val)
+
+    rows = []
+    for session_id in sorted(aggregated.keys()):
+        for model in sorted(aggregated[session_id].keys()):
+            tokens = aggregated[session_id][model]
+            rows.append({
+                "session": session_id,
+                "model": model,
+                "prompt_tokens": tokens["prompt_tokens"],
+                "cache_read_input_tokens": tokens["cache_read_input_tokens"],
+                "cache_creation_input_tokens": tokens["cache_creation_input_tokens"],
+                "completion_tokens": tokens["completion_tokens"],
+                "total_tokens": tokens["total_tokens"],
+            })
+    return rows
+
+
+def format_session_table(rows):
+    # type: (List[Dict]) -> str
+    """Render per-session *rows* as a plain-text ASCII table."""
+    if not rows:
+        return "No trace files found for the given time window."
+
+    col_keys = [
+        "session", "model", "prompt_tokens",
+        "cache_read_input_tokens", "cache_creation_input_tokens",
+        "completion_tokens", "total_tokens",
+    ]
+    widths = [len(h) for h in _SESSION_HEADERS]
+    for row in rows:
+        for i, key in enumerate(col_keys):
+            widths[i] = max(widths[i], len(str(row[key])))
+
+    sep = "+-" + "-+-".join("-" * w for w in widths) + "-+"
+    fmt = "| " + " | ".join("{{:<{}}}".format(w) for w in widths) + " |"
+
+    lines = [sep, fmt.format(*_SESSION_HEADERS), sep]
+    for row in rows:
+        lines.append(fmt.format(*[str(row[k]) for k in col_keys]))
+    lines.append(sep)
+    return "\n".join(lines)
